fix(renderer): remove redundant isize cast flagged by clippy
cf_string.char_len() already returns CFIndex (isize).

+++ rust-core/bedterm_core/src/renderer/coretext_raster.rs
@@ -139,7 +139,7 @@ impl CoreTextRasterizer {
         let cf_string = CFString::new(s);
         let range = CFRange {
             location: 0,
-            length: cf_string.char_len() as isize,
+            length: cf_string.char_len(),
         };
 
         let raw_font = unsafe {
